+++ opensyde_tool/src/opensyde_gui_elements/line_edit/C_OgeLeFilePathBase.cpp
@@ -14,6 +14,7 @@
 /* -- Includes ------------------------------------------------------------------------------------------------------ */
 #include "precomp_headers.hpp"
 
+#include <algorithm>
 #include <QFileInfo>
 #include <QDir>
 #include <QMimeData>
@@ -91,7 +92,7 @@ void C_OgeLeFilePathBase::SetPath(const QString & orc_New, const QString & orc_R
 
    // make sure last known cursor position is not out of range
    this->ms32_LastKnownCursorPos = std::min(static_cast<int32_t>(this->ms32_LastKnownCursorPos),
-                                            this->mc_Path.size());
+                                            static_cast<int32_t>(this->mc_Path.size()));
 }
 
 //----------------------------------------------------------------------------------------------------------------------
